Renamed files for `JoltObjectImpl3D`
