+++ src/objects/jolt_body_impl_3d.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "objects/jolt_collision_object_3d.hpp"
+#include "objects/jolt_object_impl_3d.hpp"
 #include "objects/jolt_physics_direct_body_state_3d.hpp"
 
 class JoltAreaImpl3D;
